+++ src/Resource.ts
@@ -13,10 +13,9 @@ export interface ResourceActionMap {
   [action: string]: ResourceAction
 }
 
-export interface ResourceActionOptions {
+export interface ShorthandResourceActionOptions {
   action: string
   property?: string
-  method: string
   path: Function | string
   onSuccess?: Function
   onError?: Function
@@ -24,8 +23,12 @@ export interface ResourceActionOptions {
   queryParams?: Boolean
 }
 
+export interface ResourceActionOptions extends ShorthandResourceActionOptions {
+  method: string
+}
+
 export interface ResourceOptions {
-  baseURL: string,
+  baseURL?: string,
   state?: Object,
   axios?: Object,
   queryParams?: Boolean